# internal/handler/http/base.go
// Code generated by meta-egg. CAREFULLY EDIT.
// WILL NOT BE replace after re-generated, unless you confirm it!
// CAREFULLY EDIT.
// Version: v1.0.4-EE
// Author: meta-egg
// Generated at: 2023-04-22 00:40
package handler

import (
	api "meta-egg-layout/api/meta_egg_layout"
	"meta-egg-layout/internal/common/cerror"
	"meta-egg-layout/internal/common/resource"
	"meta-egg-layout/internal/domain"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/go-playground/validator/v10"
	"github.com/google/wire"
	"github.com/sirupsen/logrus"
)

// ProviderSet is http handler providers.
var ProviderSet = wire.NewSet(
	NewHandler,
)

var (
	DateLayout = "20060102"
	TimeLayout = "150405"
)

type RspBase struct {
	Code    api.ErrCode `json:"code"`
	Message string      `json:"message"`
}

type RspData struct {
	RspBase
	Data interface{} `json:"data"`
}

// data如果为nil, 则响应RspBase{}; 否则响应RspData{}
func ResponseSuccess(c *gin.Context, data interface{}) {
	e := cerror.Ok()
	if data != nil {
		c.JSON(
			e.HttpStatus,
			RspData{
				RspBase: RspBase{
					Code:    e.Code,
					Message: e.Error(),
				},
				Data: data,
			},
		)
	} else {
		c.JSON(
			e.HttpStatus,
			RspBase{
				Code:    e.Code,
				Message: e.Error(),
			},
		)
	}
}

func ResponseFail(c *gin.Context, err error) {
	var (
		cErr *cerror.CustomError
		ok   bool
	)
	if cErr, ok = err.(*cerror.CustomError); !ok {
		cErr = cerror.Unknown(err.Error())
	}
	c.JSON(
		cErr.HttpStatus,
		RspBase{
			Code:    cErr.Code,
			Message: cErr.Error(),
		},
	)
}

// shouldbind and validate
func shouldBind(c *gin.Context, req interface{}) error {
	if err := c.ShouldBind(req); err != nil {
		logrus.Errorf("c.ShouldBind failed, err: %v", err)
		validationErrors, ok := err.(validator.ValidationErrors)
		if !ok {
			return cerror.InvalidArgument()
		} else {
			fields := make([]string, 0, len(validationErrors))
			for _, err := range validationErrors {
				fields = append(fields, err.Field())
			}
			return cerror.InvalidArgument(strings.Join(fields, ", "))
		}
	}
	return nil
}

type Handler struct {
	Resource      *resource.Resource
	DomainUsecase *domain.DomainUsecase

	// TODO: add your domain usecase
}

func NewHandler(
	rsrc *resource.Resource,
	domainUsecase *domain.DomainUsecase,

	// TODO: add domain usecase
) *Handler {
	return &Handler{
		Resource:      rsrc,
		DomainUsecase: domainUsecase,

		// TODO: setup domain usecase
	}
}