# internal/handler/http/user.go
// Code generated by meta-egg. CAREFULLY EDIT.
// WILL NOT BE replace after re-generated, unless you confirm it!
// CAREFULLY EDIT.
// Version: v1.0.4-EE
// Author: meta-egg
// Generated at: 2023-04-22 00:40
package handler

import (
	"meta-egg-layout/gen/model"
	"meta-egg-layout/internal/domain"
	"meta-egg-layout/internal/domain/option"

	jgstr "github.com/Jinglever/go-string"
	"github.com/gin-gonic/gin"
	log "github.com/sirupsen/logrus"
)

// 用户详情
type UserDetail struct {
	Id        uint64  `json:"id"`         //
	Name      *string `json:"name"`       // 用户名
	Gender    uint64  `json:"gender"`     // 性别
	CreatedBy *uint64 `json:"created_by"` // 创建者
	CreatedAt int64   `json:"created_at"` // 创建时间
	UpdatedBy *uint64 `json:"updated_by"` // 更新者
	UpdatedAt int64   `json:"updated_at"` // 更新时间

}

func newUserDetailFromModel(mUser *model.User) UserDetail {
	return UserDetail{
		Id:        mUser.ID,
		Name:      mUser.Name,
		Gender:    mUser.Gender,
		CreatedBy: mUser.CreatedBy,
		CreatedAt: mUser.CreatedAt.Unix(),
		UpdatedBy: mUser.UpdatedBy,
		UpdatedAt: mUser.UpdatedAt.Unix(),
	}
}

type ReqCreateUser struct {
	Name   *string `json:"name" binding:"omitempty,max=64"` // 用户名
	Gender uint64  `json:"gender" binding:"required,gte=1"` // 性别

}

//	@Id			CreateUser
//	@Tags		User
//	@Summary	创建用户
//	@Description
//	@Accept		json
//	@Produce	json
//	@Param		Authorization	header		string			true	"Bearer <jwt-token>"
//	@Param		user			body		ReqCreateUser	true	"用户"
//	@Success	200				{object}	RspData{data=UserDetail}
//	@Failure	400				{object}	RspBase
//	@Router		/v1/users [post]
func (h *Handler) CreateUser(c *gin.Context) {
	var req ReqCreateUser
	err := shouldBind(c, &req)
	if err != nil {
		ResponseFail(c, err)
		return
	}

	mUser := &model.User{
		Name:   req.Name,
		Gender: req.Gender,
	}
	err = h.DomainUsecase.CreateUser(c.Request.Context(), mUser)
	if err != nil {
		ResponseFail(c, err)
		return
	}
	ResponseSuccess(c, newUserDetailFromModel(mUser))
}

//	@Id			GetUserDetail
//	@Tags		User
//	@Summary	获取用户详情
//	@Description
//	@Accept		json
//	@Produce	json
//	@Param		Authorization	header		string	true	"Bearer <jwt-token>"
//	@Param		id				path		int		true	"用户ID"
//	@Success	200				{object}	RspData{data=UserDetail}
//	@Failure	400				{object}	RspBase
//	@Router		/v1/users/{id} [get]
func (h *Handler) GetUserDetail(c *gin.Context) {
	id := jgstr.UintVal(c.Param("id"))
	mUser, err := h.DomainUsecase.GetUserByID(c.Request.Context(), id)
	if err != nil {
		log.Errorf("domainUsecase.GetUserByID failed, err: %v", err)
		ResponseFail(c, err)
		return
	}
	ResponseSuccess(c, newUserDetailFromModel(mUser))
}

// 用户列表
type UserList struct {
	List  []*UserDetail `json:"list"`  // 用户列表
	Total int64         `json:"total"` // 总数
}

type ReqGetUserList struct {
	Page     int `form:"page" binding:"required,gte=1"`      // 页码, 从1开始
	PageSize int `form:"page_size" binding:"required,gte=1"` // 每页数量, 要求大于0
	// 筛选条件
	Gender *uint64 `form:"gender" binding:"omitempty,gte=1"` // 性别
	// 排序条件
	OrderBy   *string           `form:"order_by" binding:"omitempty,oneof=id name gender created_by created_at updated_by updated_at"` // 排序字段,可选:id|name|gender|created_by|created_at|updated_by|updated_at
	OrderType *option.OrderType `form:"order_type" binding:"omitempty,oneof=asc desc"`                                                 // 排序类型,默认desc

}

//	@Id			GetUserList
//	@Tags		User
//	@Summary	获取用户列表
//	@Description
//	@Accept		json
//	@Produce	json
//	@Param		Authorization	header	string	true	"Bearer <jwt-token>"
//	@Param		page			query	int		true	"页码, 从1开始"
//	@Param		page_size		query	int		true	"每页数量, 要求大于0"
//	@Param		gender			query	uint64	false	"性别"
//	@Param		order_by		query	string	false	"排序字段, 可选: id|name|gender|created_by|created_at|updated_by|updated_at"
//	@Param		order_type		query	string	false	"排序类型,默认desc"

//	@Success	200	{object}	RspData{data=UserList}
//	@Failure	400	{object}	RspBase
//	@Router		/v1/users [get]
func (h *Handler) GetUserList(c *gin.Context) {
	var req ReqGetUserList
	err := shouldBind(c, &req)
	if err != nil {
		ResponseFail(c, err)
		return
	}
	opt := &domain.UserListOption{
		Pagination: &option.PaginationOption{
			Page:     req.Page,
			PageSize: req.PageSize,
		},
		Filter: &domain.UserFilterOption{
			Gender: req.Gender,
		},
		Order: &option.OrderOption{
			OrderBy:   req.OrderBy,
			OrderType: req.OrderType,
		},
	}
	msUser, total, err := h.DomainUsecase.GetUserList(c.Request.Context(), opt)
	if err != nil {
		ResponseFail(c, err)
		return
	}
	list := make([]*UserDetail, 0, len(msUser))
	for _, mUser := range msUser {
		detail := newUserDetailFromModel(mUser)
		list = append(list, &detail)
	}
	ResponseSuccess(c, UserList{
		List:  list,
		Total: total,
	})
}